fix: fixed pc bug
