+++ src/common/utils/utils.ts
@@ -1,6 +1,3 @@
-import ClarityIcon from '@/ui/icons/ClarityIcon';
-import TransferIcon from '@/ui/icons/TransferIcon';
-import { ArrowsLeftRight, Cube, PhoneCall, PlusMinus, Question } from '@phosphor-icons/react';
 import BigNumber from 'bignumber.js';
 import { c32addressDecode } from 'c32check';
 import dayjs from 'dayjs';
@@ -472,59 +469,6 @@ export function promiseWrapper<T, Args extends any[]>(
       }
     });
   };
-}
-
-export function getTxTypeIcon(txType: string) {
-  switch (txType) {
-    case 'token_transfer':
-      return <TransferIcon />;
-    case 'contract_call':
-      return <PhoneCall />;
-    case 'smart_contract':
-      return <ClarityIcon />;
-    case 'tenure_change':
-      return <Cube />;
-    case 'all':
-      return <PlusMinus weight="bold" />;
-    default:
-      return <Question />;
-  }
-}
-
-export function getTxTypeLabel(txType: string) {
-  switch (txType) {
-    case 'token_transfer':
-      return 'Token Transfer';
-    case 'contract_call':
-      return 'Contract Call';
-    case 'smart_contract':
-      return 'Contract Deploy';
-    case 'tenure_change':
-      return 'Tenure Change';
-    case 'coinbase':
-      return 'Coinbase';
-    case 'all':
-      return 'Average';
-    default:
-      return 'Unknown';
-  }
-}
-
-export function getTxTypeColor(txType: string) {
-  switch (txType) {
-    case 'token_transfer':
-      return 'transactionTypes.tokenTransfer';
-    case 'contract_call':
-      return 'transactionTypes.contractCall';
-    case 'smart_contract':
-      return 'transactionTypes.contractDeploy';
-    case 'tenure_change':
-      return 'transactionTypes.tenureChange';
-    case 'coinbase':
-      return 'transactionTypes.coinbase';
-    default:
-      return 'transactionTypes.tokenTransfer';
-  }
 }
 
 function toKebabCase(str: string) {
